- Income feature preparation

+++ app/src/main/java/com/snister/carnagealpha/features/expense_tracker/data/models/IncomeDataModel.kt
@@ -11,7 +11,7 @@ data class IncomeDataModel(
 
     val incomeSourceName: String,
 
-    val incomeAmount: Double,
+    val incomeAmount: Long,
 
     val dateTime: String,
 )
