+++ app/src/main/java/com/snister/carnagealpha/features/expense_tracker/domain/entities/IncomeEntity.kt
@@ -5,6 +5,6 @@ import java.time.ZonedDateTime
 data class IncomeEntity(
     val incomeId: Int?,
     val incomeSourceName: String,
-    val incomeAmount: Double,
+    val incomeAmount: Long,
     val dateTime: ZonedDateTime,
 )
